Cleanups

+++ q3asm/cmdlib.c
@@ -255,7 +255,7 @@ void SetQdirFromPath( const char *path )
       }
 			strncpy (qdir, path, c+len+count-path);
 			qprintf ("qdir: %s\n", qdir);
-			for ( i = 0; i < strlen( qdir ); i++ )
+			for ( i = 0; i < (int)strlen( qdir ); i++ )
 			{
 				if ( qdir[i] == '\\' ) 
 					qdir[i] = '/';
@@ -268,7 +268,7 @@ void SetQdirFromPath( const char *path )
 				{
 					strncpy (gamedir, path, c+1-path);
 
-					for ( i = 0; i < strlen( gamedir ); i++ )
+					for ( i = 0; i < (int)strlen( gamedir ); i++ )
 					{
 						if ( gamedir[i] == '\\' ) 
 							gamedir[i] = '/';
